Make Authoring helpers unit testable
