+++ config/default/AuthoringService.conf.php
@@ -1,0 +1,11 @@
+<?php
+use oat\taoQtiItem\model\AuthoringService;
+
+/**
+ * Default config header created during install
+ */
+return new AuthoringService([
+    'preserveWhiteSpace' => false,
+    'formatOutput' => true,
+    'validateOnParse' => false
+]);
